Move proc.cc to use cstdint types

+++ proc.hh
@@ -1,5 +1,5 @@
 /* Honeywell Series 16 emulator
- * Copyright (C) 1997, 1998, 2005, 2010, 2011  Adrian Wise
+ * Copyright (C) 1997, 1998, 2005, 2010, 2011, 2026  Adrian Wise
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -19,6 +19,7 @@
  */
 
 #include <vector>
+#include <cstdint>
 #include "instr.hh" // To get GENERIC_GROUP_A
 #include "event_queue.hh"
 
@@ -40,7 +41,7 @@ public:
 
   void mem_access(bool p_not_pp1, bool store);
   void do_instr(bool &run, bool &monitor_flag);
-  unsigned short ea(unsigned short instr);
+  uint16_t ea(uint16_t instr);
 
   void master_clear(); // do what the master-clear button does
 
@@ -52,8 +53,8 @@ public:
    * writing to location "zero" (possibly relocated)
    * also changes the X register.
    */
-  void write(unsigned short addr, signed short data);
-  unsigned short read(unsigned short addr)
+  void write(uint16_t addr, int16_t data);
+  uint16_t read(uint16_t addr)
   {
     y = addr;
     m = core[addr & addr_mask];
@@ -65,21 +66,21 @@ public:
    *
    * Writing to X changes location "zero"
    */
-  void set_py(unsigned short n) {p = y = n;}
-  unsigned short get_p(void) {return p;}
-  void set_a(unsigned short n) {a = n;}
-  unsigned short get_a(void) {return a;}
-  void set_b(unsigned short n) {b = n;}
-  unsigned short get_b(void) {return b;}
-  void set_x(unsigned short n);
-  void set_just_x(unsigned short n);
-  unsigned short get_x(void) {return x;}
+  void set_py(uint16_t n) {p = y = n;}
+  uint16_t get_p(void) {return p;}
+  void set_a(uint16_t n) {a = n;}
+  uint16_t get_a(void) {return a;}
+  void set_b(uint16_t n) {b = n;}
+  uint16_t get_b(void) {return b;}
+  void set_x(uint16_t n);
+  void set_just_x(uint16_t n);
+  uint16_t get_x(void) {return x;}
   
   bool get_c() {return c;}
   bool get_ex() {return extend;}
   bool get_dp() {return dp;}
 
-  unsigned short get_sc() {return sc & 0x3f;}
+  uint16_t get_sc() {return sc & 0x3f;}
 
   /* 
    * Get and set sense switches
@@ -87,13 +88,13 @@ public:
   bool get_ss(int sw) {return ss[sw];}
   void set_ss(int sw, bool v) {ss[sw] = v;}
   
-  unsigned long long get_half_cycles(void){return half_cycles;}
+  uint64_t get_half_cycles(void){return half_cycles;}
 
-  void set_interrupt(unsigned short bit);
-  void clear_interrupt(unsigned short bit);
+  void set_interrupt(uint16_t bit);
+  void clear_interrupt(uint16_t bit);
   void set_rtclk(bool v);
 
-  void set_dmcreq(unsigned short bit);
+  void set_dmcreq(uint16_t bit);
   unsigned int get_dmc_dev(){return dmc_dev;}
   
   void dump_memory();
@@ -104,8 +105,8 @@ public:
   
   void start_button();
   void goto_monitor();
-  void set_limit(unsigned long long half_cycles);
-  void set_sbi(unsigned long long half_cycles);
+  void set_limit(uint64_t half_cycles);
+  void set_sbi(uint64_t half_cycles);
   
   const char *dis();
   void flush_events();
@@ -119,14 +120,14 @@ public:
   void asr_ptp_on(char *filename);
   bool special(char k);
 
-  bool optimize_io_poll(unsigned short instr);
+  bool optimize_io_poll(uint16_t instr);
 
   void abort();
 
-  void sampled_io(bool x_drlin, unsigned short x_inb){
+  void sampled_io(bool x_drlin, uint16_t x_inb){
     drlin=x_drlin; inb=x_inb;}
 
-  int get_wrt_info(unsigned short addr[2], unsigned short data[2]);
+  int get_wrt_info(uint16_t addr[2], uint16_t data[2]);
 
   void queue(unsigned long microseconds, IODEV *device, int reason)
   {
@@ -137,7 +138,7 @@ public:
 #endif
   }
   
-  void queue_hc(unsigned long long half_cycles, IODEV *device, int reason)
+  void queue_hc(uint64_t half_cycles, IODEV *device, int reason)
   {
 #ifndef RTL_SIM
     EventQueue::EventTime event_time = this->half_cycles + half_cycles;
@@ -152,31 +153,31 @@ private:
   /*
    * the following are the machine registers
    */
-  signed short a;
-  signed short b;
-  signed short x;
-  signed short m;
-  signed short op;
-  unsigned short p;
-  unsigned short y; // Address register
-  unsigned short j; // base sector relocation
+  int16_t a;
+  int16_t b;
+  int16_t x;
+  int16_t m;
+  int16_t op;
+  uint16_t p;
+  uint16_t y; // Address register
+  uint16_t j; // base sector relocation
   std::vector<bool> prt;
   
   // sampled I/O values
   bool drlin;
-  unsigned short inb;
+  uint16_t inb;
 
   // Memory write testing
   int wrts;
-  unsigned short wrt_addr[2];
-  unsigned short wrt_data[2];
+  uint16_t wrt_addr[2];
+  uint16_t wrt_data[2];
 
   /*
    * Various flags
    */
   bool pi, pi_pending;
-  unsigned short interrupts;
-  unsigned short dmc_req;
+  uint16_t interrupts;
+  uint16_t dmc_req;
   bool dmc_cyc;
   unsigned int dmc_dev;
   
@@ -198,9 +199,9 @@ private:
   bool disable_extend_pending;
   bool restrict;
   const bool extend_allowed; // i.e. whether CPU has extended addressing
-  const unsigned short addr_mask;
+  const uint16_t addr_mask;
 
-  signed short sc; // shift count
+  int16_t sc; // shift count
 
   /*
    * sense switches
@@ -210,12 +211,12 @@ private:
   /*
    * the core memory 
    */
-  signed short *core;
+  int16_t *core;
   bool *modified;
 
   bool run;     // flag to say we're still running
   bool fetched; // flag to say that an instruction has been fetched
-  unsigned short fetched_p;
+  uint16_t fetched_p;
 
   int exit_code;
   bool exit_called;
@@ -242,148 +243,148 @@ private:
   int trace_ptr;
   struct Btrace *btrace_buf;
 
-  void increment_p(unsigned short n = 1);
+  void increment_p(uint16_t n = 1);
   void write_prt(unsigned int n, uint16_t v);
 
   /* These routines are public in order that the instruction
      tables can refer to them. */
 public:
-  void unimplemented(unsigned short instr);
+  void unimplemented(uint16_t instr);
 
-  void do_CRA(unsigned short instr);
-  void do_IAB(unsigned short instr);
-  void do_IMA(unsigned short instr);
-  void do_INK(unsigned short instr);
-  void do_LDA(unsigned short instr);
-  void do_LDX(unsigned short instr);
-  void do_OTK(unsigned short instr);
-  void do_STA(unsigned short instr);
-  void do_STX(unsigned short instr);
-  void do_ACA(unsigned short instr);
-  void do_ADD(unsigned short instr);
-  void do_AOA(unsigned short instr);
-  void do_SUB(unsigned short instr);
-  void do_TCA(unsigned short instr);
-  void do_ANA(unsigned short instr);
-  void do_CSA(unsigned short instr);
-  void do_CHS(unsigned short instr);
-  void do_CMA(unsigned short instr);
-  void do_ERA(unsigned short instr);
-  void do_SSM(unsigned short instr);
-  void do_SSP(unsigned short instr);
-  void do_ALR(unsigned short instr);
-  void do_ALS(unsigned short instr);
-  void do_ARR(unsigned short instr);
-  void do_ARS(unsigned short instr);
-  void do_LGL(unsigned short instr);
-  void do_LGR(unsigned short instr);
-  void do_LLL(unsigned short instr);
-  void do_LLR(unsigned short instr);
-  void do_LLS(unsigned short instr);
-  void do_LRL(unsigned short instr);
-  void do_LRR(unsigned short instr);
-  void do_LRS(unsigned short instr);
-  void do_INA(unsigned short instr);
-  void do_OCP(unsigned short instr);
-  void do_OTA(unsigned short instr);
-  void do_SMK(unsigned short instr);
-  void do_SKS(unsigned short instr);
-  void do_CAS(unsigned short instr);
-  void do_ENB(unsigned short instr);
-  void do_HLT(unsigned short instr);
-  void do_INH(unsigned short instr);
-  void do_IRS(unsigned short instr);
-  void do_JMP(unsigned short instr);
-  void do_JST(unsigned short instr);
-  void do_NOP(unsigned short instr);
-  void do_RCB(unsigned short instr);
-  void do_SCB(unsigned short instr);
-  void do_SKP(unsigned short instr);
-  void do_SLN(unsigned short instr);
-  void do_SLZ(unsigned short instr);
-  void do_SMI(unsigned short instr);
-  void do_SNZ(unsigned short instr);
-  void do_SPL(unsigned short instr);
-  void do_SR1(unsigned short instr);
-  void do_SR2(unsigned short instr);
-  void do_SR3(unsigned short instr);
-  void do_SR4(unsigned short instr);
-  void do_SRC(unsigned short instr);
-  void do_SS1(unsigned short instr);
-  void do_SS2(unsigned short instr);
-  void do_SS3(unsigned short instr);
-  void do_SS4(unsigned short instr);
-  void do_SSC(unsigned short instr);
-  void do_SSR(unsigned short instr);
-  void do_SSS(unsigned short instr);
-  void do_SZE(unsigned short instr);
-  void do_CAL(unsigned short instr);
-  void do_CAR(unsigned short instr);
-  void do_ICA(unsigned short instr);
-  void do_ICL(unsigned short instr);
-  void do_ICR(unsigned short instr);
+  void do_CRA(uint16_t instr);
+  void do_IAB(uint16_t instr);
+  void do_IMA(uint16_t instr);
+  void do_INK(uint16_t instr);
+  void do_LDA(uint16_t instr);
+  void do_LDX(uint16_t instr);
+  void do_OTK(uint16_t instr);
+  void do_STA(uint16_t instr);
+  void do_STX(uint16_t instr);
+  void do_ACA(uint16_t instr);
+  void do_ADD(uint16_t instr);
+  void do_AOA(uint16_t instr);
+  void do_SUB(uint16_t instr);
+  void do_TCA(uint16_t instr);
+  void do_ANA(uint16_t instr);
+  void do_CSA(uint16_t instr);
+  void do_CHS(uint16_t instr);
+  void do_CMA(uint16_t instr);
+  void do_ERA(uint16_t instr);
+  void do_SSM(uint16_t instr);
+  void do_SSP(uint16_t instr);
+  void do_ALR(uint16_t instr);
+  void do_ALS(uint16_t instr);
+  void do_ARR(uint16_t instr);
+  void do_ARS(uint16_t instr);
+  void do_LGL(uint16_t instr);
+  void do_LGR(uint16_t instr);
+  void do_LLL(uint16_t instr);
+  void do_LLR(uint16_t instr);
+  void do_LLS(uint16_t instr);
+  void do_LRL(uint16_t instr);
+  void do_LRR(uint16_t instr);
+  void do_LRS(uint16_t instr);
+  void do_INA(uint16_t instr);
+  void do_OCP(uint16_t instr);
+  void do_OTA(uint16_t instr);
+  void do_SMK(uint16_t instr);
+  void do_SKS(uint16_t instr);
+  void do_CAS(uint16_t instr);
+  void do_ENB(uint16_t instr);
+  void do_HLT(uint16_t instr);
+  void do_INH(uint16_t instr);
+  void do_IRS(uint16_t instr);
+  void do_JMP(uint16_t instr);
+  void do_JST(uint16_t instr);
+  void do_NOP(uint16_t instr);
+  void do_RCB(uint16_t instr);
+  void do_SCB(uint16_t instr);
+  void do_SKP(uint16_t instr);
+  void do_SLN(uint16_t instr);
+  void do_SLZ(uint16_t instr);
+  void do_SMI(uint16_t instr);
+  void do_SNZ(uint16_t instr);
+  void do_SPL(uint16_t instr);
+  void do_SR1(uint16_t instr);
+  void do_SR2(uint16_t instr);
+  void do_SR3(uint16_t instr);
+  void do_SR4(uint16_t instr);
+  void do_SRC(uint16_t instr);
+  void do_SS1(uint16_t instr);
+  void do_SS2(uint16_t instr);
+  void do_SS3(uint16_t instr);
+  void do_SS4(uint16_t instr);
+  void do_SSC(uint16_t instr);
+  void do_SSR(uint16_t instr);
+  void do_SSS(uint16_t instr);
+  void do_SZE(uint16_t instr);
+  void do_CAL(uint16_t instr);
+  void do_CAR(uint16_t instr);
+  void do_ICA(uint16_t instr);
+  void do_ICL(uint16_t instr);
+  void do_ICR(uint16_t instr);
 
 #if ((!defined(GENERIC_GROUP_A)) || defined(TEST_GENERIC_GROUP_A))
   /*
    * NPL group A
    */
-  void do_ad1(unsigned short instr);
-  void do_ad1_15(unsigned short instr);
-  void do_adc(unsigned short instr);
-  void do_adc_15(unsigned short instr);
-  void do_cm1(unsigned short instr);
-  void do_ltr(unsigned short instr);
-  void do_btr(unsigned short instr);
-  void do_btl(unsigned short instr);
-  void do_rtl(unsigned short instr);
-  void do_rcb_ssp(unsigned short instr);
-  void do_cpy(unsigned short instr);
-  void do_btb(unsigned short instr);
-  void do_bcl(unsigned short instr);
-  void do_bcr(unsigned short instr);
-  void do_ld1(unsigned short instr);
-  void do_isg(unsigned short instr);
-  void do_cma_aca(unsigned short instr);
-  void do_cma_aca_c(unsigned short instr);
-  void do_a2a(unsigned short instr);
-  void do_a2c(unsigned short instr);
-  void do_ics(unsigned short instr);
-  void do_scb_a2a(unsigned short instr);
-  void do_scb_aoa(unsigned short instr);
-  void do_a2c_scb(unsigned short instr);
-  void do_aca_scb(unsigned short instr);
-  void do_icr_scb(unsigned short instr);
-  void do_rtl_scb(unsigned short instr);
-  void do_btb_scb(unsigned short instr);
-  void do_noa(unsigned short instr);
+  void do_ad1(uint16_t instr);
+  void do_ad1_15(uint16_t instr);
+  void do_adc(uint16_t instr);
+  void do_adc_15(uint16_t instr);
+  void do_cm1(uint16_t instr);
+  void do_ltr(uint16_t instr);
+  void do_btr(uint16_t instr);
+  void do_btl(uint16_t instr);
+  void do_rtl(uint16_t instr);
+  void do_rcb_ssp(uint16_t instr);
+  void do_cpy(uint16_t instr);
+  void do_btb(uint16_t instr);
+  void do_bcl(uint16_t instr);
+  void do_bcr(uint16_t instr);
+  void do_ld1(uint16_t instr);
+  void do_isg(uint16_t instr);
+  void do_cma_aca(uint16_t instr);
+  void do_cma_aca_c(uint16_t instr);
+  void do_a2a(uint16_t instr);
+  void do_a2c(uint16_t instr);
+  void do_ics(uint16_t instr);
+  void do_scb_a2a(uint16_t instr);
+  void do_scb_aoa(uint16_t instr);
+  void do_a2c_scb(uint16_t instr);
+  void do_aca_scb(uint16_t instr);
+  void do_icr_scb(uint16_t instr);
+  void do_rtl_scb(uint16_t instr);
+  void do_btb_scb(uint16_t instr);
+  void do_noa(uint16_t instr);
 #endif
 
-  void do_EXA(unsigned short instr);
-  void do_DXA(unsigned short instr);
+  void do_EXA(uint16_t instr);
+  void do_DXA(uint16_t instr);
 
-  void do_ERM(unsigned short instr);
+  void do_ERM(uint16_t instr);
 
-  void do_RMP(unsigned short instr);
+  void do_RMP(uint16_t instr);
 
-  void do_DBL(unsigned short instr);
-  void do_DIV(unsigned short instr);
-  void do_MPY(unsigned short instr);
-  void do_NRM(unsigned short instr);
-  void do_SCA(unsigned short instr);
-  void do_SGL(unsigned short instr);
-  void do_iab_sca(unsigned short instr);
+  void do_DBL(uint16_t instr);
+  void do_DIV(uint16_t instr);
+  void do_MPY(uint16_t instr);
+  void do_NRM(uint16_t instr);
+  void do_SCA(uint16_t instr);
+  void do_SGL(uint16_t instr);
+  void do_iab_sca(uint16_t instr);
 
-  void generic_shift(unsigned short instr);
-  void generic_skip(unsigned short instr);
+  void generic_shift(uint16_t instr);
+  void generic_skip(uint16_t instr);
 #ifdef TEST_GENERIC_SKIP
   void test_generic_skip();
 #endif
-  void generic_group_A(unsigned short instr);
+  void generic_group_A(uint16_t instr);
 #ifdef TEST_GENERIC_GROUP_A
   void test_generic_group_A();
 #endif
 
-  static signed short ex_sc(unsigned short instr);
+  static int16_t ex_sc(uint16_t instr);
 
 };
